requirements.txt
Django
django-modeltranslation
django-parler
django-parler-rest
django-rosetta
djangorestframework
